+++ v3/crates/jsonapi/tests/jsonapi_golden_tests.rs
@@ -214,6 +214,7 @@ fn get_metadata_resolve_configuration() -> metadata_resolve::configuration::Conf
     let unstable_features = metadata_resolve::configuration::UnstableFeatures {
         enable_ndc_v02_support: false,
         enable_aggregation_predicates: false,
+        enable_pre_route_plugins: false,
     };
 
     metadata_resolve::configuration::Configuration { unstable_features }