+++ v3/crates/metadata-resolve/src/types/warning.rs
@@ -2,7 +2,8 @@ use open_dds::flags;
 
 use crate::stages::{
     aggregate_boolean_expressions, aggregates, boolean_expressions, commands, data_connectors,
-    models, models_graphql, object_boolean_expressions, scalar_boolean_expressions, scalar_types,
+    models, models_graphql, object_boolean_expressions, plugins, scalar_boolean_expressions,
+    scalar_types,
 };
 
 use super::error::ShouldBeAnError;
@@ -35,6 +36,8 @@ pub enum Warning {
     AggregateExpressionIssue(#[from] aggregates::AggregateExpressionIssue),
     #[error("{0}")]
     ScalarTypesIssue(#[from] scalar_types::ScalarTypesIssue),
+    #[error("{0}")]
+    PluginIssue(#[from] plugins::PluginIssue),
 }
 
 impl ShouldBeAnError for Warning {